requirements.txt
django
dotenv
djangorestframework
django-cors-headers
firebase-admin
google.generativeai
groq
numpy